requirements.txt
djangorestframework
markdown
django-filter
django-extra-fields